+++ src/lib/components/table/transactions/TransactionsTableHeader.tsx
@@ -5,11 +5,13 @@ import { TableHeader } from "../tableComponents";
 
 export const TransactionsTableHeader = ({
   templateColumns,
+  showSuccess,
   showRelations,
   showTimestamp,
   showAction,
 }: {
   templateColumns: GridProps["templateColumns"];
+  showSuccess: boolean;
   showRelations: boolean;
   showTimestamp: boolean;
   showAction: boolean;
@@ -17,7 +19,7 @@ export const TransactionsTableHeader = ({
   <Grid templateColumns={templateColumns} minW="min-content">
     <TableHeader />
     <TableHeader>Transaction Hash</TableHeader>
-    <TableHeader />
+    {showSuccess && <TableHeader />}
     <TableHeader>Messages</TableHeader>
     {showRelations && <TableHeader>Relations</TableHeader>}
     <TableHeader>Sender</TableHeader>