+++ src/lib/services/tx/index.ts
@@ -6,13 +6,13 @@ import type {
   AccountTxsResponse,
   BlockTxsResponse,
   TxData,
-  TxsByAddressResponseLcd,
   TxsResponse,
-} from "../types/tx";
+} from "../types";
 import {
   CELATONE_QUERY_KEYS,
   useBaseApiRoute,
   useCelatoneApp,
+  useCurrentChain,
   useInitia,
   useLcdEndpoint,
   useMoveConfig,
@@ -25,9 +25,15 @@ import type {
   BechAddr20,
   BechAddr32,
   Option,
+  Transaction,
   TxFilters,
 } from "lib/types";
-import { extractTxLogs, isTxHash, snakeToCamel } from "lib/utils";
+import {
+  convertAccountPubkeyToAccountAddress,
+  extractTxLogs,
+  isTxHash,
+  snakeToCamel,
+} from "lib/utils";
 
 import {
   getTxData,
@@ -238,11 +244,31 @@ export const useTxsByContractAddressLcd = (
   address: BechAddr32,
   limit: number,
   offset: number,
-  options: UseQueryOptions<TxsByAddressResponseLcd> = {}
+  options: UseQueryOptions<TxsResponse> = {}
 ) => {
   const endpoint = useLcdEndpoint();
+  const {
+    chain: { bech32_prefix: prefix },
+  } = useCurrentChain();
 
-  return useQuery<TxsByAddressResponseLcd>(
+  const queryfn = useCallback(
+    () =>
+      getTxsByContractAddressLcd(endpoint, address, limit, offset).then(
+        (txs) => ({
+          items: txs.items.map<Transaction>((tx) => ({
+            ...tx,
+            sender: convertAccountPubkeyToAccountAddress(
+              tx.signerPubkey,
+              prefix
+            ),
+          })),
+          total: txs.total,
+        })
+      ),
+    [address, endpoint, limit, offset, prefix]
+  );
+
+  return useQuery<TxsResponse>(
     [
       CELATONE_QUERY_KEYS.TXS_BY_CONTRACT_ADDRESS_LCD,
       endpoint,
@@ -250,7 +276,7 @@ export const useTxsByContractAddressLcd = (
       limit,
       offset,
     ],
-    async () => getTxsByContractAddressLcd(endpoint, address, limit, offset),
+    queryfn,
     { retry: 1, refetchOnWindowFocus: false, ...options }
   );
 };
@@ -284,34 +310,56 @@ export const useTxsByAddressLcd = (
   search: string,
   limit: number,
   offset: number,
-  options: Pick<UseQueryOptions<TxsResponse>, "onSuccess"> = {}
+  options: UseQueryOptions<TxsResponse> = {}
 ) => {
   const endpoint = useLcdEndpoint();
   const { validateContractAddress } = useValidateAddress();
+  const {
+    chain: { bech32_prefix: prefix },
+  } = useCurrentChain();
 
-  const queryfn = useCallback(() => {
-    if (isTxHash(search)) return getTxsByHashLcd(endpoint, search);
+  const queryfn = useCallback(async () => {
+    const txs = await (async () => {
+      if (isTxHash(search)) return getTxsByHashLcd(endpoint, search);
 
-    if (!validateContractAddress(search))
-      return getTxsByContractAddressLcd(
-        endpoint,
-        search as BechAddr32,
-        offset,
-        limit
-      );
+      if (!validateContractAddress(search))
+        return getTxsByContractAddressLcd(
+          endpoint,
+          search as BechAddr32,
+          limit,
+          offset
+        );
 
-    if (!address) throw new Error("address is undefined (useTxsByAddressLcd)");
-    return getTxsByAccountAddressLcd(endpoint, address, limit, offset);
-  }, [address, endpoint, limit, offset, search, validateContractAddress]);
+      if (!address)
+        throw new Error("address is undefined (useTxsByAddressLcd)");
+      return getTxsByAccountAddressLcd(endpoint, address, limit, offset);
+    })();
 
-  return useQuery(
+    return {
+      items: txs.items.map<Transaction>((tx) => ({
+        ...tx,
+        sender: convertAccountPubkeyToAccountAddress(tx.signerPubkey, prefix),
+      })),
+      total: txs.total,
+    };
+  }, [
+    address,
+    endpoint,
+    limit,
+    offset,
+    prefix,
+    search,
+    validateContractAddress,
+  ]);
+
+  return useQuery<TxsResponse>(
     [
       CELATONE_QUERY_KEYS.TXS_BY_ADDRESS_LCD,
       endpoint,
       address,
       search,
-      offset,
       limit,
+      offset,
     ],
     queryfn,
     { ...options, retry: 1, refetchOnWindowFocus: false }