+++ src/lib/utils/tx/createTxHash.ts
@@ -1,0 +1,5 @@
+import { sha256 } from "@cosmjs/crypto";
+import { toHex } from "@cosmjs/encoding";
+
+export const createTxHash = (txRaw: string) =>
+  toHex(sha256(Buffer.from(txRaw, "base64"))).toUpperCase();